+++ handlers/test_handler.js
@@ -1,0 +1,31 @@
+/**
+ * Functions for handling request for testing purposes
+ * Calls sqlQuery function from utils and gets data from database
+ */
+
+const utils = require('../setup/utils');
+
+module.exports = {
+  getBetyg: async (req, res) => {
+    let result = [];
+
+    result = await utils.sqlQuery(
+      'SELECT BETYGSVARDE AS betyg, COUNT(BETYGSVARDE) AS antal FROM io_studieresultat WHERE UTBILDNING_KOD="TNG033" GROUP BY BETYGSVARDE'
+    );
+
+    res.status(200).send({
+      data: result,
+    });
+  },
+
+  getAvbrott: async (req, res) => {
+    let result = [];
+
+    result = await utils.sqlQuery(
+      'SELECT UTBILDNING_KOD as kurskod, COUNT(AVBROTT_UTBILDNING) as avbrott FROM io_studieresultat WHERE YTTERSTA_KURSPAKETERING_SV = "Civilingenjörsprogram i medieteknik" GROUP BY UTBILDNING_KOD ORDER BY avbrott DESC LIMIT 20'
+    );
+    res.status(200).send({
+      data: result,
+    });
+  },
+};
